update: Default to "Examples" rather than "Playground"
If JavaScript is disabled then the playground can not be shown, so a
table of examples is shown to the user instead. However, the link at
the top of the page always said "Playground", even if it was
unavailable.

The page has changed to default to the text (and link) for "Examples".
A small bit of JavaScript will run, as part of the playground script,
to update the link and text to "Playground" if it's available.